automatically attach decompileJar as sources for the remapped minecraft jar
