+++ paperweight-core/src/main/kotlin/taskcontainers/AllTasks.kt
@@ -64,6 +64,8 @@ open class AllTasks(
 
         inputJar.set(copyResources.flatMap { it.outputJar })
         libraries.from(downloadMcLibraries.map { it.outputDir.asFileTree })
+
+        outputJar.set(cache.resolve(FINAL_DECOMPILE_JAR))
     }
 
     val applyApiPatches by tasks.registering<ApplyGitPatches> {
