+++ paperweight-lib/src/main/kotlin/util/constants/constants.kt
@@ -86,6 +86,7 @@ private const val CONFIGURATION_CACHE = "$PAPER_PATH/configurationCache"
 private const val TASK_CACHE = "$PAPER_PATH/taskCache"
 
 const val FINAL_REMAPPED_JAR = "$TASK_CACHE/minecraft.jar"
+const val FINAL_DECOMPILE_JAR = "$TASK_CACHE/decompileJar.jar"
 
 const val MC_DEV_SOURCES_DIR = "$PAPER_PATH/mc-dev-sources"
